+++ internal/web/urlvals/params/page_params.go
@@ -1,7 +1,7 @@
 package params
 
 type PageParams struct {
-	Limit  *int32 `page:"limit,default=10"`
-	Offset *int32 `page:"offset"`
-	Total  *int32
+	Limit  *int64 `page:"limit,default=10"`
+	Offset *int64 `page:"offset"`
+	Total  *int64
 }
